py-worker/requirements.txt
# minimal runtime
pika==1.3.2
psycopg2-binary==2.9.9
huggingface_hub==0.24.6
scikit-learn==1.2.1
numpy==1.26.4
joblib==1.3.2
